リポジトリをソートして表示するようにした
- ファイルをソートする関数と同じファイルにまとめた
