+++ frontend/src/app/repositories/[id]/page.tsx
@@ -4,7 +4,7 @@ import { auth } from '@/auth';
 import RepositoryDetail from '@/components/repositories/RepositoryDetail';
 import { FileItem, Repository } from '@/types';
 import { fetcher } from '@/utils/fetcher';
-import { sortFileItems } from '@/utils/sort-file-items';
+import { sortFileItems } from '@/utils/sort';
 
 type RepositoryDetailPageProps = {
   params: Promise<{ id: string }>;
