+++ frontend/src/components/repositories/FileTreeItem.tsx
@@ -4,7 +4,7 @@ import { Check, ChevronDown, ChevronRight, File, Folder } from 'lucide-react';
 import { useState } from 'react';
 
 import { FileItem } from '@/types';
-import { sortFileItems } from '@/utils/sort-file-items';
+import { sortFileItems } from '@/utils/sort';
 
 type FileTreeItemProps = {
   fileItem: FileItem;
